Added functionality to update profile page

+++ backend/src/utils/s3.js
@@ -9,34 +9,36 @@ let s3 = new S3({
 });
 
 function upload(file) {
-  console.log("filedetails", file);
-  const filestream = fs.createReadStream(file.path);
-  const params = {
-    Body: filestream,
-    Key: "profile-pics/" + file.originalname,
-    Bucket: process.env.bucketName,
-  };
   return new Promise((resolve, reject) => {
-    s3.upload(params, function (err, data) {
-      if (err) {
-        console.log("Error", err);
-        reject(err);
-      }
-      if (data) {
-        console.log("Upload Success", data.Location);
-        fs.unlink(file.path, (err) => {
-          if (err != null) {
-            console.error("error occured during deleting file", err);
-          }
-        });
-        resolve(data);
-      }
-    });
+    if (file != null) {
+      console.log("filedetails", file);
+      const filestream = fs.createReadStream(file.path);
+      const params = {
+        Body: filestream,
+        Key: "profile-pics/" + file.originalname,
+        Bucket: process.env.bucketName,
+      };
+      s3.upload(params, function (err, data) {
+        if (err) {
+          console.log("Error", err);
+          reject(err);
+        }
+        if (data) {
+          console.log("Upload Success", data.Location);
+          fs.unlink(file.path, (err) => {
+            if (err != null) {
+              console.error("error occured during deleting file", err);
+            }
+          });
+          resolve(data);
+        }
+      });
+    }else{
+      resolve("No file")
+    }
   });
 }
 
-function download() {
-
-}
+function download() {}
 
 module.exports = { upload, download };
